rearranging classsifier
